+++ docs/changelog.md
@@ -20,6 +20,9 @@ permalink: /changelog/
     into a new component `support-images`, which provides helpers for handling images. `AndroidIconDecoder` and `IconDecoder`
     are renamed to `AndroidImageDecoder` and `ImageDecoder` in `support-images`.
 
+* **browser-tabstray**
+  * When tabs are updated, the `BrowserTabsTray` now scrolls to the selected tab.
+
 # 42.0.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v41.0.0...42.0.0)